NEC2: Add flushing around file saves.
The extra flushing is to get the data on disk before xnec2c opens the file when
optimizing.  If it gets a partial copy it may crash.  There is a bugfix pending
for xnec2c as well that ensures the file is >0 bytes before loading.  Hopefully
the probelem is solved between this commit and the commit in xnec2c.